70% done with build, working on Router

+++ rabbit-api/db/migrate/20201223010630_create_habits.rb
@@ -4,7 +4,7 @@ class CreateHabits < ActiveRecord::Migration[6.0]
       t.string :name
       t.string :goal
       t.date :start_day
-      t.float :streak
+      t.integer :streak
       t.integer :user_id
 
       t.timestamps
